+++ app/src/main/java/cn/spacexc/wearbili/dataclass/comment/list/FansDetail.kt
@@ -37,5 +37,5 @@ data class FansDetail(
     @SerializedName("score")
     val score: Int,
     @SerializedName("uid")
-    val uid: Int
+    val uid: Long
 )
