+++ app/src/main/java/cn/spacexc/wearbili/dataclass/comment/list/Pendant.kt
@@ -15,5 +15,5 @@ data class Pendant(
     @SerializedName("name")
     val name: String,
     @SerializedName("pid")
-    val pid: Int
+    val pid: Long
 )
